Separate back into cwrapper and cwrapper_ndarrays
