+++ pyccel/stdlib/cwrapper/cwrapper.h
@@ -6,7 +6,6 @@
 # include <complex.h>
 # include <stdint.h>
 # include <stdbool.h>
-# include "ndarrays.h"
 
 # define NO_IMPORT_ARRAY
 # define PY_ARRAY_UNIQUE_SYMBOL CWRAPPER_ARRAY_API
@@ -15,17 +14,7 @@
 # define NO_TYPE_CHECK -1
 # define NO_ORDER_CHECK -1
 
-
-
-t_ndarray	pyarray_to_c_ndarray(PyArrayObject *o);
-t_ndarray	pyarray_to_f_ndarray(PyArrayObject *o);
-
-
-/* arrays checkers and helpers */
-bool	pyarray_check(PyArrayObject *o, int dtype, int rank, int flag);
-
-void    *nd_data(t_ndarray *a);
-int     nd_ndim(t_ndarray *a, int n);
+extern const char* dataTypes[17];
 
 
 /* casting python object to c type */
