+++ pyccel/stdlib/cwrapper/cwrapper_ndarrays.c
@@ -51,8 +51,136 @@ static int64_t     *_numpy_to_ndarray_shape(npy_intp  *np_shape, int nd)
     for (int i = 0; i < nd; i++)
         nd_shape[i] = (int64_t) np_shape[i];
     return nd_shape;
-
 }
+
+/*
+ * Function: _check_pyarray_dtype
+ * --------------------
+ * Check Python Object DataType:
+ *
+ * 	Parameters	:
+ *		a 	  : python array object
+ *      dtype : desired data type enum
+ * 	Returns		:
+ *		return true if no error occurred otherwise it will return false
+ *      and raise TypeError exception
+ * reference of the used c/python api function
+ * -------------------------------------------
+ * https://numpy.org/doc/stable/reference/c-api/array.html#c.PyArray_TYPE
+ */
+bool	check_pyarray_dtype(PyArrayObject *a, int dtype)
+{
+	int current_dtype;
+
+	if (dtype == NO_TYPE_CHECK)
+		return true;
+
+	current_dtype = PyArray_TYPE(a);
+	if (current_dtype != dtype)
+	{
+		PyErr_Format(PyExc_TypeError,
+			"argument dtype must be %s, not %s",
+			dataTypes[dtype],
+			dataTypes[current_dtype]);
+		return false;
+	}
+
+	return true;
+}
+
+/*
+ * Function: _check_pyarray_rank
+ * --------------------
+ * Check Python Object Rank:
+ *
+ * 	Parameters	:
+ *		a 	  : python array object
+ *      rank  : desired rank
+ * 	Returns		:
+ *		return true if no error occurred otherwise it will return false
+ *      and raise TypeError exception
+ * reference of the used c/python api function
+ * -------------------------------------------
+ * https://numpy.org/doc/stable/reference/c-api/array.html#c.PyArray_NDIM
+ */
+static bool _check_pyarray_rank(PyArrayObject *a, int rank)
+{
+	int	current_rank;
+
+	current_rank = PyArray_NDIM(a);
+	if (current_rank != rank)
+	{
+		PyErr_Format(PyExc_TypeError, "argument rank must be %d, not %d",
+			rank,
+			current_rank);
+		return false;
+	}
+
+	return true;
+}
+
+/*
+ * Function: _check_pyarray_order
+ * --------------------
+ * Check Python Object Order:
+ *
+ * 	Parameters	:
+ *		a 	  : python array object
+ *      flag  : desired order
+ * 	Returns		:
+ *		return true if no error occurred otherwise it will return false
+ *      and raise NotImplementedError exception
+ * reference of the used c/python api function
+ * -------------------------------------------
+ * https://numpy.org/doc/stable/reference/c-api/array.html#c.PyArray_CHKFLAGS
+ */
+static bool _check_pyarray_order(PyArrayObject *a, int flag)
+{
+	char	order;
+
+	if (flag == NO_ORDER_CHECK)
+		return true;
+
+	if (!PyArray_CHKFLAGS(a, flag))
+	{
+		order = flag == NPY_ARRAY_C_CONTIGUOUS ? 'C' : 'F';
+		PyErr_Format(PyExc_NotImplementedError,
+			"argument does not have the expected ordering (%c)", order);
+		return false;
+	}
+
+	return true;
+}
+
+
+/*
+ * Function: _check_pyarray_type
+ * --------------------
+ * Check if Python Object is ArrayType:
+ *
+ * 	Parameters	:
+ *		a 	  : python array object
+ *
+ * 	Returns		:
+ *		return true if no error occurred otherwise it will return false
+ *      and raise TypeError exception
+ * reference of the used c/python api function
+ * -------------------------------------------
+ * https://numpy.org/doc/stable/reference/c-api/array.html#c.PyArray_Check
+ */
+static bool _check_pyarray_type(PyObject *a)
+{
+	if (!PyArray_Check(a))
+	{
+		PyErr_Format(PyExc_TypeError,
+			"argument must be numpy.ndarray, not %s",
+			 a == Py_None ? "None" : Py_TYPE(a)->tp_name);
+		return false;
+	}
+
+	return true;
+}
+
 
 /* converting numpy array to c nd array*/
 t_ndarray	pyarray_to_c_ndarray(PyArrayObject *a)
@@ -71,4 +199,78 @@ t_ndarray	pyarray_to_c_ndarray(PyArrayObject *a)
 	array.is_view     = 1;
 
 	return array;
+}
+
+
+/*
+ * Function: pyarray_check
+ * --------------------
+ * Check Python Object (DataType, Rank, Order):
+ *
+ * 	Parameters	:
+ *		a 	  : python array object
+ *      dtype : desired data type enum
+ *		rank  : desired rank
+ *		flag  : desired order flag
+ * 	Returns		:
+ *		return true if no error occurred otherwise it will return false
+ */
+bool	pyarray_check(PyArrayObject *o, int dtype, int rank, int flag)
+{
+	if (!_check_pyarray_type((PyObject *)o)) return false;
+
+	// check array element type / rank / order
+	if(!check_pyarray_dtype(o, dtype)) return false;
+
+	if(!_check_pyarray_rank(o, rank)) return false;
+
+	if(rank > 1 && !_check_pyarray_order(o, flag)) return false;
+
+	return true;
+}
+
+
+/*
+ * Function: nd_ndim
+ * --------------------
+ * Return the shape in the n dimension.
+ *
+ * 	Parameters	:
+ *		a 	  : python array object
+ *      index : dimension index
+ * 	Returns		:
+ *		return 0 if object is NULL or shape at indexed dimension
+ * reference of the used c/numpy api function
+ * -------------------------------------------
+ * https://numpy.org/doc/1.17/reference/c-api.array.html#c.PyArray_DIM
+ */
+int     nd_ndim(t_ndarray *a, int n)
+{
+	if (a == NULL)
+		return 0;
+	
+	return a->shape[n];
+}
+
+
+/*
+ * Function: nd_data
+ * --------------------
+ * Return data pointed by array
+ *
+ * 	Parameters	:
+ *		a 	  : python array object
+ *      index : dimension index
+ * 	Returns		:
+ *		return NULL if object is NULL or the data of the array
+ * reference of the used c/numpy api function
+ * -------------------------------------------
+ * https://numpy.org/doc/1.17/reference/c-api.array.html#c.PyArray_DIM
+ */
+void    *nd_data(t_ndarray *a)
+{
+	if (a == NULL)
+		return NULL;
+	
+	return a->data;
 }
