+++ pyccel/stdlib/cwrapper/cwrapper_ndarrays.h
@@ -34,4 +34,14 @@
  */
 t_ndarray	pyarray_to_c_ndarray(PyArrayObject *o);
 
+t_ndarray	pyarray_to_c_ndarray(PyArrayObject *o);
+t_ndarray	pyarray_to_f_ndarray(PyArrayObject *o);
+
+
+/* arrays checkers and helpers */
+bool	pyarray_check(PyArrayObject *o, int dtype, int rank, int flag);
+
+void    *nd_data(t_ndarray *a);
+int     nd_ndim(t_ndarray *a, int n);
+
 #endif
